add cf argument for SstFileWriter constructor (#22)
