+++ src/rocksdb.rs
@@ -16,7 +16,7 @@
 use crocksdb_ffi::{self, DBWriteBatch, DBCFHandle, DBInstance, DBBackupEngine,
                    DBStatisticsTickerType, DBStatisticsHistogramType};
 use libc::{self, c_int, c_void, size_t};
-use rocksdb_options::{Options, ReadOptions, UnsafeSnap, WriteOptions, FlushOptions,
+use rocksdb_options::{Options, ReadOptions, UnsafeSnap, WriteOptions, FlushOptions, EnvOptions,
                       RestoreOptions, IngestExternalFileOptions, HistogramData};
 use std::{fs, ptr, slice};
 use std::collections::BTreeMap;
@@ -1301,6 +1301,61 @@ impl Drop for BackupEngine {
     }
 }
 
+/// SstFileWriter is used to create sst files that can be added to database later
+/// All keys in files generated by SstFileWriter will have sequence number = 0
+pub struct SstFileWriter {
+    inner: *mut crocksdb_ffi::SstFileWriter,
+}
+
+unsafe impl Send for SstFileWriter {}
+
+impl SstFileWriter {
+    pub fn new(env_opt: &EnvOptions, opt: &Options, cf: &CFHandle) -> SstFileWriter {
+        unsafe {
+            SstFileWriter {
+                inner: crocksdb_ffi::crocksdb_sstfilewriter_create(env_opt.inner,
+                                                                   opt.inner,
+                                                                   cf.inner),
+            }
+        }
+    }
+
+    /// Prepare SstFileWriter to write into file located at "file_path".
+    pub fn open(&mut self, name: &str) -> Result<(), String> {
+        let path = match CString::new(name.to_owned()) {
+            Err(e) => return Err(format!("invalid path {}: {:?}", name, e)),
+            Ok(p) => p,
+        };
+        unsafe { Ok(ffi_try!(crocksdb_sstfilewriter_open(self.inner, path.as_ptr()))) }
+    }
+
+    /// Add key, value to currently opened file
+    /// REQUIRES: key is after any previously added key according to comparator.
+    pub fn add(&mut self, key: &[u8], val: &[u8]) -> Result<(), String> {
+        unsafe {
+            ffi_try!(crocksdb_sstfilewriter_add(self.inner,
+                                                key.as_ptr(),
+                                                key.len(),
+                                                val.as_ptr(),
+                                                val.len()));
+            Ok(())
+        }
+    }
+
+    /// Finalize writing to sst file and close file.
+    pub fn finish(&mut self) -> Result<(), String> {
+        unsafe {
+            ffi_try!(crocksdb_sstfilewriter_finish(self.inner));
+            Ok(())
+        }
+    }
+}
+
+impl Drop for SstFileWriter {
+    fn drop(&mut self) {
+        unsafe { crocksdb_ffi::crocksdb_sstfilewriter_destroy(self.inner) }
+    }
+}
 
 #[cfg(test)]
 mod test {
